We were not using the most efficient cryptographic algorithm. We were using DES, but the AES implementation is much more efficient. As an added bonus, we are more secure because DES keys are 56 bits and we use 128 bit keys for AES.
The digest output stream was updating the signature state every time write() was called. This was particularly inefficient. Now, we send the data through a DigestOutputStream and we compute the signature only once.

Going from DES to AES uncovered a bug in the data protection service. The data protection did not support keys that were bigger than DES keys.