There is a problem that occured with thhe old technique when a message from A to B can't be recieved and the protection level message going back, that would fix the problem cannot be received either.
I therefore modified the ProtectionLevelMessage approach to use
exceptions.  A new aspect, RemotePolicyExceptionAspect, catches
exceptions (IncorrectProtectionException and
SenderUsingInvalidCertException) that indicate that the receiver
rejected a message.  The receiver then tries to correct the error and
retry.

The one ProtectionLevelMessage I kept was StopSigningMessage.  This
message is not urgent and will not cause the society to lock up if it
is not received.