Added the ability to skip committing policies to particular enclaves. On the TIC this allows me to avoid committing the restrictive policies to the 1-UA enclave which can't handle it.
I also added a fix to the commit policies method because when it was
told that the policy change was a delta it did a set instead.  In the
non-restrictive policy tests on the TIC, we are only using the set
policies version of the commit method.

I have also changed the InstallURPolicies to be a policy delta.  This
means that I need to install the boot policies before I take this action.